backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication
djangorestframework-simplejwt==5.3.1
django-allauth==0.57.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# WebSocket support
channels==4.0.0
channels-redis==4.1.0

# Kafka
confluent-kafka==2.3.0

# Redis
redis==5.0.1
hiredis==2.3.2

# Celery (async tasks)
celery==5.3.4

# AI/ML
openai==1.6.1

# Utilities
python-dotenv==1.0.0
Pillow==10.1.0
python-decouple==3.8

# Monitoring & Logging
sentry-sdk==1.39.2

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
flake8==7.0.0
black==23.12.1
isort==5.13.2

# WSGI/ASGI Servers
gunicorn==21.2.0
uvicorn==0.25.0
daphne==4.0.0

# QR Code generation
qrcode==7.4.2
